project/requirements.txt
bleach==1.5.0
cycler==0.10.0
enum34==1.1.6
html5lib==0.9999999
Markdown==2.6.9
matplotlib==2.1.0
numpy==1.13.3
protobuf==3.5.0.post1
pygame==1.9.3
pyparsing==2.2.0
python-dateutil==2.6.1
python-interface==1.4.0
pytz==2017.3
PyYAML==3.12
six==1.11.0
Werkzeug==0.12.2
zope.event==4.3.0
zope.interface==4.4.3
zope.schema==4.5.0
